requirements.txt
streamlit
langchain
langchain-groq
langchain-pinecone
langchain-community
langchain-huggingface
pinecone
plotly
pandas
python-dotenv
tiktoken